telegram-service/requirements.txt
python-telegram-bot==13.15
redis==5.0.1
kafka-python 